add 2 other examples gmail and mailslurp
