+++ .github/workflows/main.yml
@@ -32,9 +32,6 @@ jobs:
       - name: Checkout 🛎️
         uses: actions/checkout@v2
 
-      - name: Download Config from S3 📥
-        run: aws s3 sync s3://cypress-for-everything ./
-
       - name: Get yarn cache directory path
         id: yarn-cache-dir-path
         run: echo "::set-output name=dir::$(yarn cache dir)"
@@ -61,12 +58,6 @@ jobs:
 
       - name: Install dependencies 🔧
         run: yarn install --frozen-lockfile --prefer-offline
-
-      - name: Refresh Gmail Access Token 🔄
-        run: yarn refreshGmailToken
-
-      - name: Upload Config to S3 📤
-        run: aws s3 sync ./cypress/plugins s3://cypress-for-everything/cypress/plugins --exclude "*.ts" --include "*.json"
 
       - name: Create Matrix Examples ⚡
         id: matrix
@@ -99,6 +90,7 @@ jobs:
         id: cypress-test
         run: |
           cd examples/${{ matrix.example }}
+          yarn postsetup || echo "no postsetup"
           yarn test
 
       - name: Upload Video 📹
