+++ .gitignore
@@ -110,7 +110,3 @@ report.xml
 # Ignore Cypress Temp files from examples
 examples/*/cypress/videos/*.*
 examples/*/cypress/screenshots/*.*
-
-cypress/plugins/credentials.json
-cypress/plugins/token.json
-cypress/fixtures/*.*
