+++ cypress/plugins/index.ts
@@ -1,4 +1,3 @@
-import { gmailCheck, GmailCheckParam } from "./gmailCheck";
 import {
   structuredData,
   StructuredDataParam,
@@ -18,12 +17,6 @@ module.exports = (on, config) => {
       param: StructuredDataParam
     ): Promise<StructuredDataResult> {
       return await structuredData(param);
-    },
-  });
-
-  on("task", {
-    async gmailCheck(args: GmailCheckParam) {
-      return await gmailCheck(args);
     },
   });
   return config;
