--- cypress/plugins/refreshGmailToken.ts
@@ -1,10 +1,0 @@
-const path = require("path");
-const gmail = require("gmail-tester");
-
-(async function () {
-  console.log("Refreshed the gmail token!");
-  await gmail.refresh_access_token(
-    path.resolve(__dirname, "credentials.json"),
-    path.resolve(__dirname, "token.json")
-  );
-})();
