+++ examples/gmail/.env.example
@@ -1,0 +1,17 @@
+
+# Example for Gmail Account "mariorossifake@gmail.com"
+
+## the domain of your gmail/gsuite account
+CYPRESS_EMAIL_PRE=mariorossifake
+
+## the domain of your gmail/gsuite account
+CYPRESS_EMAIL_POST=gmail.com
+
+
+# Example for Gsuite Account "mariorossifake@mariorossiwebsite.com"
+
+## the domain of your gmail/gsuite account
+CYPRESS_EMAIL_PRE=mariorossifake
+
+## the domain of your gmail/gsuite account
+CYPRESS_EMAIL_POST=mariorossiwebsite.com
