+++ examples/gmail/.gitignore
@@ -1,0 +1,2 @@
+cypress/plugins/credentials.json
+cypress/plugins/token.json
