+++ examples/gmail/cypress.json
@@ -1,7 +1,6 @@
 {
   "fixturesFolder": false,
   "supportFile": false,
-  "pluginsFile": false,
   "reporter": "junit",
   "reporterOptions": {
     "mochaFile": "../../report/[hash].xml",
