+++ examples/gmail/cypress/plugins/index.ts
@@ -1,0 +1,17 @@
+import { gmailCheck, GmailCheckParam } from "./gmailCheck";
+
+require("ts-node").register({
+  transpileOnly: true,
+});
+
+/**
+ * @type {Cypress.PluginConfig}
+ */
+module.exports = (on, config) => {
+  on("task", {
+    async gmailCheck(args: GmailCheckParam) {
+      return await gmailCheck(args);
+    },
+  });
+  return config;
+};
