+++ examples/gmail/package.json
@@ -3,8 +3,9 @@
   "private": true,
   "version": "0.0.0",
   "scripts": {
-    "dev": "cypress open",
-    "test": "echo 'gmail'"
+    "dev": "doppler run -- cypress open",
+    "test": "cypress run",
+    "postsetup": "yarn ts-node ./cypress/plugins/refreshGmailToken"
   },
   "dependencies": {
     "cypress": "^8.3.1",
