+++ examples/mailslurp/.env.example
@@ -1,0 +1,1 @@
+CYPRESS_MAILSLURP_API_KEY=https://app.mailslurp.com/sign-up/
