+++ examples/mailslurp/cypress.json
@@ -1,6 +1,5 @@
 {
   "fixturesFolder": false,
-  "supportFile": false,
   "pluginsFile": false,
   "reporter": "junit",
   "reporterOptions": {
