+++ examples/mailslurp/cypress/support/index.js
@@ -13,10 +13,4 @@
 // https://on.cypress.io/configuration
 // ***********************************************************
 
-// Import commands.js using ES2015 syntax:
-import "./commands";
-
-// Alternatively you can use CommonJS syntax:
-// require('./commands')
-
 import "cypress-mailslurp";
