+++ examples/mailslurp/package.json
@@ -3,8 +3,8 @@
   "private": true,
   "version": "0.0.0",
   "scripts": {
-    "dev": "cypress open",
-    "test": "echo 'mailslurp'"
+    "dev": "doppler run -- cypress open",
+    "test": "cypress run"
   },
   "dependencies": {
     "cypress": "^8.3.1",
